+++ js-miniapp-bridge/src/index.ts
@@ -31,7 +31,7 @@ import {
   MAAnalyticsActionType,
   MAAnalyticsEventType,
 } from './types/analytics/analytics';
-import { UniversalBridgeToHostInfo } from './types/universal-bridge';
+import { UniversalBridgeInfo } from './types/universal-bridge';
 
 import {
   AuthorizationFailureError,
@@ -110,5 +110,5 @@ export {
   MAAnalyticsInfo,
   MAAnalyticsActionType,
   MAAnalyticsEventType,
-  UniversalBridgeToHostInfo,
+  UniversalBridgeInfo,
 };