+++ js-miniapp-bridge/src/types/universal-bridge.ts
@@ -1,7 +1,7 @@
 /**
  * Model that will be sent to Host app through Universal bridge
  */
-export interface UniversalBridgeToHostInfo {
+export interface UniversalBridgeInfo {
   key?: string;
   value?: string;
   description?: string;