+++ js-miniapp-sdk/CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 1.18.0 (2023-XX-XX)
 - **Feature:** Added new interface `isDarkMode()`. Interface to know if the Hostapp/Device is set to use the Dark 
 - **Feature:** Added new interface `sendAnalytics(analyticsInfo: MAAnalyticsInfo)`. This interface will be used to send analytics to host app.
-- **Feature:** Added Universal Bridge related interface e.g. `sendInfoToHostapp` to send `UniversalBridgeToHostInfo` to the HostApp. 
+- **Feature:** Added Universal Bridge related interface e.g. `sendInfoToHostapp` to send `UniversalBridgeInfo` to the HostApp. 
 
 
 ### 1.17.0 (2023-03-30)